action과 reducer를 reduxt toolkit을 이용하여 통합합니다.
